Split leveldb from plyvel tests, removed print statement
